refactor feedback on cloze screen
