+++ src/components/ClozeFeedback/types.ts
@@ -1,0 +1,14 @@
+export interface ClozeFeedbackProps {
+  isCorrect: boolean;
+  correctWord: string;
+  userAnswer: string;
+  translation: string;
+  sentence: string;
+  points: number;
+  newMastery: number;
+  previousMastery: number;
+  onNext: () => void;
+  onAddToAnki: () => void;
+  isAddingToAnki?: boolean;
+  ankiAdded?: boolean;
+}
